+++ go-controller/pkg/clustermanager/id_allocator.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"sync"
 
-	bitmapallocator "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/ipallocator/allocator"
+	bitmapallocator "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ipallocator/allocator"
 )
 
 const (
